requirements.txt
datasets
pandas
transformers
tensorflow
numpy
